refactor .netframework tests
