+++ csharp/test/Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2/.gitignore
