+++ csharp/test/Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2/NuGet.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+<solution>
+    <add key="disableSourceControlIntegration" value="true" />
+</solution>
+<packageSources>
+ <add key="NuGet Official" value="https://api.nuget.org/v3/index.json" />
+ <add key="BuildArtifacts" value="E:\nugets" />
+</packageSources>
+</configuration>
