+++ csharp/test/Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.AI.MachineLearning" version="1.3.1-dev-20200626-0616-f44313c3d" targetFramework="net472" />
+  <package id="Microsoft.AI.MachineLearning" version="1.4.0-dev-20200728-0957-6cb93ccb5" targetFramework="net472" />
   <package id="Microsoft.Windows.CsWinRT" version="0.1.0-prerelease.200629.3" targetFramework="net472" />
   <package id="Microsoft.Windows.SDK.NET" version="10.0.18362.3-preview" targetFramework="net472" />
 </packages>
