+++ tools/ci_build/github/azure-pipelines/nuget/templates/windowsai.yml
@@ -346,26 +346,34 @@ jobs:
       workingDirectory: $(Build.ArtifactStagingDirectory)\merged
 
   - task: PowerShell@2
-    displayName: 'NuGet Tests: Restore Windows.AI.MachineLearning Nuget Package (C#)'
+    displayName: 'NuGet Tests: Build Tests (CppWinRT)'
+    inputs:
+      targetType: 'inline'
+      script: |
+        msbuild /p:Platform=x86 Microsoft.AI.MachineLearning.Tests.vcxproj
+        msbuild /p:Platform=x64 Microsoft.AI.MachineLearning.Tests.vcxproj
+      workingDirectory: $(Build.SourcesDirectory)\csharp\test\Microsoft.AI.MachineLearning.Tests
+
+  - task: PowerShell@2
+    displayName: 'NuGet Tests: Fix Nuget Package references (.NET FRAMEWORK 4.7.2)'
     inputs:
       targetType: 'inline'
       script: |
         Add-Type -AssemblyName "System.IO.Compression.FileSystem"
 
+        $src_root_dir = $Env:BUILD_SOURCESDIRECTORY;
+        $artifacts_staging_dir = $Env:BUILD_ARTIFACTSTAGINGDIRECTORY;
+        $merged_nuget_path = [System.IO.Path]::Combine($artifacts_staging_dir, 'merged')
         $nupkgs = (Get-ChildItem -Filter Microsoft.AI.MachineLearning*)
         $merged_nuget_package = $nupkgs[0]
         $merged_nuget_package_name = $merged_nuget_package.Name
         $matched_name = ($merged_nuget_package_name -match "Microsoft.AI.MachineLearning.(?<version>.*).nupkg")
         $package_version = $matches['version']
 
-        $src_root_dir = $Env:BUILD_SOURCESDIRECTORY;
-        $artifacts_staging_dir = $Env:BUILD_ARTIFACTSTAGINGDIRECTORY;
-        $merged_nuget_path = [System.IO.Path]::Combine($artifacts_staging_dir, 'merged')
-        $src_dir = [System.IO.Path]::Combine($src_root_dir, 'csharp', 'test', 'Microsoft.AI.MachineLearning.Tests.CSharp')
+        $src_dir = [System.IO.Path]::Combine($src_root_dir, 'csharp', 'test', 'Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2')
+        $input_csproj = [System.IO.Path]::Combine($src_dir, 'Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2.csproj.pp')
+        $csproj = [System.IO.Path]::Combine($src_dir, 'Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2.csproj')
         $packages_dir = [System.IO.Path]::Combine($src_dir, 'packages')
-
-        $input_csproj = [System.IO.Path]::Combine($src_dir, 'Microsoft.AI.MachineLearning.Tests.CSharp.csproj.pp')
-        $csproj = [System.IO.Path]::Combine($src_dir, 'Microsoft.AI.MachineLearning.Tests.CSharp.csproj')
         $input_csproj_content = Get-Content -Path $input_csproj
         $csproj_content = $input_csproj_content -replace '\[PackageVersion\]', $package_version
         Set-Content -Path $csproj -Value $csproj_content
@@ -380,23 +388,13 @@ jobs:
       workingDirectory: $(Build.ArtifactStagingDirectory)\merged
 
   - task: PowerShell@2
-    displayName: 'NuGet Tests: Build Tests (CppWinRT)'
-    inputs:
-      targetType: 'inline'
-      script: |
-        msbuild /p:Platform=x86 Microsoft.AI.MachineLearning.Tests.vcxproj
-        msbuild /p:Platform=x64 Microsoft.AI.MachineLearning.Tests.vcxproj
-      workingDirectory: $(Build.SourcesDirectory)\csharp\test\Microsoft.AI.MachineLearning.Tests
-
-  - task: PowerShell@2
     displayName: 'NuGet Tests: Build Tests (C#)'
     inputs:
       targetType: 'inline'
-      script: |
-        msbuild Microsoft.AI.MachineLearning.Tests.CSharp.csproj /p:Platform=AnyCpu
-        msbuild Microsoft.AI.MachineLearning.Tests.CSharp.csproj /p:Platform=x64
-      workingDirectory: $(Build.SourcesDirectory)\csharp\test\Microsoft.AI.MachineLearning.Tests.CSharp
-
+      script: |        
+        msbuild Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2.csproj /p:Platform=AnyCpu
+        msbuild Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2.csproj /p:Platform=x64
+      workingDirectory: $(Build.SourcesDirectory)\csharp\test\Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2
 
   - task: PowerShell@2
     displayName: 'NuGet Tests: Fix Nuget Package references (.NET CORE 3.0)'
@@ -405,18 +403,16 @@ jobs:
       script: |
         Add-Type -AssemblyName "System.IO.Compression.FileSystem"
 
+        $src_root_dir = $Env:BUILD_SOURCESDIRECTORY;
+        $artifacts_staging_dir = $Env:BUILD_ARTIFACTSTAGINGDIRECTORY;
+        $merged_nuget_path = [System.IO.Path]::Combine($artifacts_staging_dir, 'merged')
         $nupkgs = (Get-ChildItem -Filter Microsoft.AI.MachineLearning*)
         $merged_nuget_package = $nupkgs[0]
         $merged_nuget_package_name = $merged_nuget_package.Name
         $matched_name = ($merged_nuget_package_name -match "Microsoft.AI.MachineLearning.(?<version>.*).nupkg")
         $package_version = $matches['version']
 
-        $src_root_dir = $Env:BUILD_SOURCESDIRECTORY;
-        $artifacts_staging_dir = $Env:BUILD_ARTIFACTSTAGINGDIRECTORY;
-        $merged_nuget_path = [System.IO.Path]::Combine($artifacts_staging_dir, 'merged')
         $src_dir = [System.IO.Path]::Combine($src_root_dir, 'csharp', 'test', 'Microsoft.AI.MachineLearning.Tests.DotNetCore3_0')
-        $packages_dir = [System.IO.Path]::Combine($src_dir, 'packages')
-
         $input_csproj = [System.IO.Path]::Combine($src_dir, 'Microsoft.AI.MachineLearning.Tests.DotNetCore3_0.csproj.pp')
         $csproj = [System.IO.Path]::Combine($src_dir, 'Microsoft.AI.MachineLearning.Tests.DotNetCore3_0.csproj')
         $input_csproj_content = Get-Content -Path $input_csproj
@@ -461,12 +457,12 @@ jobs:
         popd
         
         Write-Host "Run Microsoft.AI.MachineLearning CSharp Tests (AnyCpu)"
-        pushd .\Microsoft.AI.MachineLearning.Tests.CSharp\bin\Debug
-        .\Microsoft.AI.MachineLearning.Tests.CSharp.exe
+        pushd .\Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2\bin\Debug
+        .\Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2.exe
         popd
         Write-Host "Run Microsoft.AI.MachineLearning CSharp Tests (x64)"
-        pushd .\Microsoft.AI.MachineLearning.Tests.CSharp\bin\x64\Debug
-        .\Microsoft.AI.MachineLearning.Tests.CSharp.exe
+        pushd .\Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2\bin\x64\Debug
+        .\Microsoft.AI.MachineLearning.Tests.DotNetFramework4_7_2.exe
         popd
         Write-Host "Done!"
       workingDirectory: $(Build.SourcesDirectory)\csharp\test
